Update methods for deposit

+++ FinanceAnalytic/Transactions/TransferTransaction.cs
@@ -6,14 +6,14 @@ namespace FinanceAnalytic
 {
     class TransferTransaction : ITransactions
     {
-        public TransferTransaction(double sum, Category category, DateTime today)
+        public TransferTransaction(decimal sum, Category category, DateTime today)
         {
             Sum = sum;
             Category = category.Name;
             Date = today;
         }
 
-        public double Sum { get; }
+        public decimal Sum { get; }
         public DateTime Date { get; }
 
         public string Category { get; set; }
